fixing new codes

+++ src/main/java/com/scuticommerce/model/cart/Item.java
@@ -2,11 +2,14 @@ package com.scuticommerce.model.cart;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Data
 public class Item {
 
     private String id;
     private String itemId;
     private Integer quantity;
-    private Float unitPrice;
+    private BigDecimal unitPrice;
 }
